# Update montgomery multiplication to use s2n-bignum's verified scalar bignum functions