--- a/crypto/fipsmodule/bn/montgomery.c
+++ b/crypto/fipsmodule/bn/montgomery.c
@@ -139,8 +139,13 @@ OPENSSL_INLINE int montgomery_use_s2n_bignum(unsigned int num) {
   // (4) Temporary buffer's size (t and mulres) used in
   //     montgomery_s2n_bignum_mul_mont does not exceed
   //     BN_MONTGOMERY_MAX_WORDS.
+  assert(S2NBIGNUM_KSQR_16_32_TEMP_NWORDS <= S2NBIGNUM_KMUL_32_64_TEMP_NWORDS &&
+         S2NBIGNUM_KSQR_32_64_TEMP_NWORDS <= S2NBIGNUM_KMUL_32_64_TEMP_NWORDS &&
+         S2NBIGNUM_KMUL_16_32_TEMP_NWORDS <= S2NBIGNUM_KMUL_32_64_TEMP_NWORDS);
+  const uint64_t temp_buffer_nwords =
+      S2NBIGNUM_KMUL_32_64_TEMP_NWORDS + 2 * (uint64_t)num;
   return !CRYPTO_is_ARMv8_wide_multiplier_capable() && (num % 8 == 0) &&
-         BN_BITS2 == 64 && (2 * (uint64_t)num + 96) <= BN_MONTGOMERY_MAX_WORDS;
+         BN_BITS2 == 64 && temp_buffer_nwords <= BN_MONTGOMERY_MAX_WORDS;
 }
 
 #else
@@ -465,15 +470,15 @@ static void montgomery_s2n_bignum_mul_mont(BN_ULONG *rp, const BN_ULONG *ap,
 
 #if defined(BN_MONTGOMERY_USE_S2N_BIGNUM)
 
-  // t is a temporary buffer used by big-int multiplication.
+  // t is a temporary buffer used by Karatsuba multiplication.
   // bignum_kmul_32_64 requires 96 words.
-  uint64_t t[96];
+  uint64_t t[S2NBIGNUM_KMUL_32_64_TEMP_NWORDS];
   // mulres is the output buffer of big-int multiplication.
-  // If BN_MONTGOMERY_MAX_WORDS - 96 is larger than num*2, its low num*2
-  // elements are used.
+  // If BN_MONTGOMERY_MAX_WORDS - S2NBIGNUM_KMUL_32_64_TEMP_NWORDS is larger
+  // than num*2, its low num*2 elements are used.
   // It is montgomery_use_s2n_bignum() that checks whether num*2 fits in the
   // size of mulres array.
-  uint64_t mulres[BN_MONTGOMERY_MAX_WORDS - 96];
+  uint64_t mulres[BN_MONTGOMERY_MAX_WORDS - S2NBIGNUM_KMUL_32_64_TEMP_NWORDS];
 
   // Given m the prime number stored at np, m * w = -1 mod 2^64.
   uint64_t w = n0[0];