@dkostic 2023-08-07
nice improvements, can we get perf numbers for GV3 and possibly Apple M1/M2?