@aqjune-aws 2023-08-07
@dkostic For microarchitectures supporting fast enough multiplications (such as Graviton 2 and M1), the current version of s2n-bignum didn't bring effective performance improvement. To selectively apply s2n-bignum to Graviton 2, this patch invokes `CRYPTO_is_ARMv8_wide_multiplier_capable()` and runs the s2n-bignum functions only when wide multipliers are not capable.
This is the relevant part in arm_arch.h:
```c
// The Neoverse V1 and Apple M1 micro-architectures are detected to enable
// high unrolling factor of AES-GCM and other algorithms that leverage a
// wide crypto pipeline and fast multiplier.
#define ARMV8_NEOVERSE_V1 (1 << 12)
#define ARMV8_APPLE_M1 (1 << 13)
```