@aqjune-aws 2023-08-10
From the CI failure, it seems in ARM the s2n-bignum assembly files are not linked to executables if `MY_ASSEMBLER_IS_TOO_OLD_FOR_AVX` is set.

Relevant CMakeLists.txt lines: https://github.com/aws/aws-lc/blob/main/crypto/fipsmodule/CMakeLists.txt#L170-L171

Should we update the CMakeLists.txt lines in this PR?